requirements.txt
beautifulsoup4==4.12.3
lxml==5.3.0
matplotlib==3.9.3
networkx==3.4.2
numpy==2.1.3
pandas==2.2.3
Pillow==11.0.0
prettytable==3.12.0
